# PERF: performance improvement in MultiIndex.sortlevel