-------------------------------------------------------------------------------
    Test name                                    | head[ms] | base[ms] |  ratio   |
    -------------------------------------------------------------------------------
    multiindex_sortlevel_int64                   | 664.1977 | 2443.2590 |   0.2718 |
    -------------------------------------------------------------------------------
    Test name                                    | head[ms] | base[ms] |  ratio   |
    -------------------------------------------------------------------------------

    Ratio < 1.0 means the target commit is faster then the baseline.
    Seed used: 1234

    Target [dd31bae] : performance improvement in MultiIndex.sortlevel
    Base   [c4a996a] : Merge pull request #9411 from dashesy/add_sql_test

    test sql table name